// src/types/general.d.ts
export enum Provider {
  NONE = "none", // initial value for the abstract adapter, don't use this one
  LOCAL = "local", // local testing adapter
  GCS = "gcs", // Google Cloud Storage
  GS = "gs", // Google Cloud Storage
  S3 = "s3", // Amazon S3
  AWS = "aws", // Amazon S3
  AZURE = "azure", // Azure Storage Blob
  B2 = "b2", // BackBlaze B2 using native API with AdapterBackblazeB2
  BACKBLAZE = "b2", // BackBlaze B2 using native API with AdapterBackblazeB2
  B2_S3 = "b2-s3", // Backblaze B2 using S3 API with AdapterAmazonS3
  BACKBLAZE_S3 = "b2-s3", // Backblaze B2 using S3 API with AdapterAmazonS3
  MINIO = "minio", // Minio using native API with AdapterMinio
  MINIO_S3 = "minio-s3", // Minio using S3 API with AdapterAmazonS3
  CUBBIT = "cubbit", // Cubbit uses S3 API with AdapterAmazonS3
  R2 = "r2", // Cloudflare R2 uses S3 API with AdapterAmazonS3
  CLOUDFLARE = "r2", // Cloudflare R2 uses S3 API with AdapterAmazonS3
}
export interface AdapterConfig {
  bucketName?: string;
  [id: string]: any; // eslint-disable-line
}

export interface StorageAdapterConfig extends AdapterConfig {
  provider: string;
}
export interface Options {
  [id: string]: any; // eslint-disable-line
}